Add a branch name check workflow
This adds a workflow that can be used in combination with the
allowed-prefixes.yml config to enforce standard branch naming
conventions.

This checks both for the standard prefix defined in the config
and also checks that the rest of the branch name is composed
only of lowercase letters, numbers, and/or dashes.

repo usage example:
```
jobs:
  branch-name-check:
    uses: city-of-saint-louis/.github/workflows/branch-name-check.yml@main
```