Permit exemptions for branch names
The branch name workflow should be able to make exceptions for
branch name checks.

This will prevent false notifications when it is necessary to
push a branch with a different naming convention (like main or
ghpages).

This changes the workflow call needed in individual repos:
```
name: Branch Name Check

on:
  push:
    branches:
      - '**'

jobs:
  branch-name-check:
    uses: city-of-saint-louis/.github/workflow/branch-name-check.yml@main
    with:
      exempt_branches: "main,develop,release"
```