Use HEAD_REF for the target branch name
The child workflow trigger event needed to be changed
to run on pull requests instead of push.

This means that we need to use
`BRANCH_NAME="${GITHUB_HEAD_REF:-${GITHUB_REF#refs/heads/}}`
instead of
`BRANCH_NAME="${GITHUB_REF#refs/heads/}"` to get the right name
for the source branch.